features: count score, navbar multiplayer-dropdown, bugs fixed: creating only one player in create_game and join_game (ajax multiple fireing problem), editing riddle problem fixed

+++ ajax/delete.php
@@ -2,7 +2,7 @@
 
 session_start();
 
-include ("../../projekt/notLoggedRedirect.php");
+include("../../projekt/notLoggedRedirect.php");
 include("../admin/includes/Riddle.php");
 
 
@@ -14,9 +14,12 @@ if (isset($_POST['id'])) {
 }
 
 $deleteRiddle = new Riddle();
-$deleteRiddle->id=$id;
+$deleteRiddle->id = $id;
+$riddle=Riddle::findById($deleteRiddle->id);
 
-if ($deleteRiddle->delete()) {
-    echo 1;
-} else echo 0;
+if ($riddle->in_match==0) {
+    if ($deleteRiddle->delete()) {
+        echo 1;
+    } else echo "There was a problem with deleting riddle.";
+}else echo "You can't delete riddle that takes or took part in the game.";
 exit;
